EndpointAndIndexesConfigurator_initIndex_optimization - revert old command, add new command config change
