+++ src/Resources/config/pimcore/config.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 use CIHub\Bundle\SimpleRESTAdapterBundle\Messenger\DeleteIndexElementMessage;
 use CIHub\Bundle\SimpleRESTAdapterBundle\Messenger\UpdateIndexElementMessage;
 use CIHub\Bundle\SimpleRESTAdapterBundle\Messenger\RebuildIndexElementMessage;
+use CIHub\Bundle\SimpleRESTAdapterBundle\Messenger\RebuildUpdateIndexElementMessage;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
@@ -34,6 +35,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
                 DeleteIndexElementMessage::class => 'datahub_es_index_queue',
                 UpdateIndexElementMessage::class => 'datahub_es_index_queue',
                 RebuildIndexElementMessage::class => 'datahub_es_index_queue',
+                RebuildUpdateIndexElementMessage::class => 'datahub_es_index_queue',
             ],
         ],
     ]);
